# back/aventurier.openapi.yml
openapi: 3.1.1
info:
  title: Aventurier API
  version: 1.0.0
  description: API pour gérer des aventuriers
...
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/aventuriers:
    get:
      summary: Liste tous les aventuriers
      description: Retourne la liste des aventuriers. Rôle requis -> VIEWER
      operationId: getAllAventuriers
      tags:
        - Aventurier
      responses:
        '200':
          description: Liste des aventuriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Aventurier'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    post:
      summary: Crée un nouvel aventurier
      description: Crée un nouvel aventurier. Rôle requis -> ADMIN
      operationId: createAventurier
      tags:
        - Aventurier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AventurierCreate'
      responses:
        '201':
          description: Aventurier créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aventurier'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/aventuriers/{id}:
    get:
      summary: Récupère un aventurier par son identifiant
      description: Retourne le détail d'un aventurier. Rôle requis -> VIEWER
      operationId: getAventurierById
      tags:
        - Aventurier
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant de l'aventurier
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Aventurier trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aventurier'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    patch:
      summary: Met à jour partiellement un aventurier
      description: Met à jour un aventurier. Rôle requis -> ADMIN
      operationId: updateAventurier
      tags:
        - Aventurier
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant de l'aventurier
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AventurierUpdate'
      responses:
        '200':
          description: Aventurier mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aventurier'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Conflict - Violation de règle métier (ex. niveau)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      summary: Supprime un aventurier
      description: Supprime un aventurier. Rôle requis -> ADMIN
      operationId: deleteAventurier
      tags:
        - Aventurier
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant de l'aventurier
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Aventurier supprimé
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/aventuriers/{id}/competences:
    get:
      summary: Liste les compétences d'un aventurier
      description: Retourne les compétences possédées par un aventurier. Rôle requis -> VIEWER
      operationId: getCompetencesByAventurierId
      tags:
        - Aventurier
        - Competence
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant de l'aventurier
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Liste des compétences de l'aventurier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Competence'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/aventuriers/{id}/competences/disponibles:
    get:
      summary: Liste les compétences disponibles pour un aventurier
      description: Retourne les compétences accessibles (prérequis remplis et non déjà possédées). Rôle requis -> VIEWER
      operationId: getCompetencesDisponiblesByAventurierId
      tags:
        - Aventurier
        - Competence
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant de l'aventurier
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Compétences acquérables et bloquées pour l'aventurier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetencesDisponibles'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/aventuriers/{id}/competences/{cId}:
    post:
      summary: Attribue une compétence à un aventurier
      description: Attribue une compétence si les prérequis sont respectés. Rôle requis -> ADMIN
      operationId: addCompetenceToAventurier
      tags:
        - Aventurier
        - Competence
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant de l'aventurier
          schema:
            type: integer
            format: int64
        - name: cId
          in: path
          required: true
          description: Identifiant de la compétence
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Compétence attribuée
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Conflict - Violation d'invariant métier (prérequis non remplis)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      summary: Retire une compétence d'un aventurier
      description: Retire une compétence si elle n'est pas prérequis d'une autre déjà possédée. Rôle requis -> ADMIN
      operationId: removeCompetenceFromAventurier
      tags:
        - Aventurier
        - Competence
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant de l'aventurier
          schema:
            type: integer
            format: int64
        - name: cId
          in: path
          required: true
          description: Identifiant de la compétence
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Compétence retirée
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Conflict - Violation d'invariant métier (compétence requise par une autre possédée)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/competences/{id}/aventuriers:
    get:
      summary: Liste les aventuriers possédant une compétence
      description: Retourne possesseurs et aventuriers éligibles pour une compétence. Rôle requis -> VIEWER
      operationId: getAventuriersByCompetenceId
      tags:
        - Aventurier
        - Competence
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant de la compétence
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Possesseurs et aventuriers éligibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AventuriersByCompetence'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Detail
      properties:
        type:
          type: string
          format: uri
          description: URI identifiant le type de problème
        title:
          type: string
          description: Résumé court du problème
        status:
          type: integer
          format: int32
          description: Code HTTP
        detail:
          type: string
          description: Explication détaillée du problème
        instance:
          type: string
          format: uri
          description: URI identifiant l'occurrence du problème
    Classe:
      type: string
      description: Classe de l'aventurier
      enum:
        - MAGE
        - GUERRIER
        - ARCHER
        - RÔDEUR
        - PRÊTRE
        - CHANTEUR
        - NÉGATEUR
        - SORCIER
        - PALADIN
        - BARD
    Caracteristique:
      type: string
      description: Caractéristique d'un aventurier
      enum:
        - PHYSIQUE
        - MENTAL
        - PERCEPTION
    CaracteristiqueMin:
      type: object
      description: Prérequis sur une caractéristique (valeur minimale)
      properties:
        caracteristique:
          $ref: '#/components/schemas/Caracteristique'
        valeur:
          type: integer
          format: int32
      required:
        - caracteristique
        - valeur
    PrerequisCompetence:
      type: object
      description: Prérequis d'une compétence (tous optionnels)
      properties:
        classeRequise:
          $ref: '#/components/schemas/Classe'
        niveauMinimum:
          type: integer
          format: int32
          minimum: 1
        caracteristiquesMin:
          $ref: '#/components/schemas/CaracteristiqueMin'
        competencesRequises:
          type: array
          description: Liste d'identifiants de compétences requises (déjà connues)
          items:
            type: string
            format: uuid
    CompetenceRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identifiant unique de la compétence (UUID)
      required:
        - id
    CompetenceEditableAttributes:
      type: object
      description: Champs éditables d'une compétence
      properties:
        nom:
          type: string
          description: Nom unique de la compétence
        description:
          type: string
          description: Description de la compétence
        prerequis:
          $ref: '#/components/schemas/PrerequisCompetence'
    Competence:
      allOf:
        - $ref: '#/components/schemas/CompetenceRef'
        - $ref: '#/components/schemas/CompetenceEditableAttributes'
        - type: object
          required:
            - id
            - nom
...
    AventurierRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identifiant unique de l'aventurier (auto-incrémenté)
      required:
        - id
    AventurierEditableAttributes:
      type: object
      description: Champs éditables d'un aventurier (tous optionnels pris isolément)
      properties:
        nom:
          type: string
          description: Nom de l'aventurier
        description:
          type: string
          description: Description de l'aventurier
        physique:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          description: Note de physique sur 100
        mental:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          description: Note de mental sur 100
        perception:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          description: Note de perception sur 100
        niveau:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          description: Niveau global sur 100
        classe:
          $ref: '#/components/schemas/Classe'
    Aventurier:
      allOf:
        - $ref: '#/components/schemas/AventurierRef'
        - $ref: '#/components/schemas/AventurierEditableAttributes'
        - type: object
          required:
            - id
            - nom
            - physique
            - mental
            - perception
            - niveau
            - classe
    AventurierCreate:
      description: Données pour créer un aventurier. Le niveau est ignoré (toujours initialisé à 1).
      allOf:
        - $ref: '#/components/schemas/AventurierEditableAttributes'
        - type: object
          required:
            - nom
            - physique
            - mental
            - perception
            - classe
    AventurierUpdate:
      description: Champs modifiables d'un aventurier (tous optionnels). Le niveau ne peut jamais descendre et ne peut monter que de 1 à la fois.
      allOf:
        - $ref: '#/components/schemas/AventurierEditableAttributes'
...
    CompetencesBloqueesItem:
      type: object
      description: Compétence non acquise avec les prérequis manquants
      properties:
        competence:
          $ref: '#/components/schemas/Competence'
        prerequisManquants:
          type: array
          description: Liste lisible des prérequis non satisfaits
          items:
            type: string
    CompetencesDisponibles:
      type: object
      description: Compétences acquérables et bloquées pour un aventurier
      properties:
        acquerables:
          type: array
          description: Compétences non encore acquises dont tous les prérequis sont satisfaits
          items:
            $ref: '#/components/schemas/Competence'
        bloquees:
          type: array
          description: Compétences non acquises avec leurs prérequis manquants
          items:
            $ref: '#/components/schemas/CompetencesBloqueesItem'
    AventuriersByCompetence:
      type: object
      description: Possesseurs et aventuriers éligibles pour une compétence
      properties:
        possesseurs:
          type: array
          description: Aventuriers qui possèdent déjà cette compétence
          items:
            $ref: '#/components/schemas/Aventurier'
        eligibles:
          type: array
          description: Aventuriers qui pourraient acquérir cette compétence maintenant
          items:
            $ref: '#/components/schemas/Aventurier'